# tests/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelNet",
    "description": "校园网加密流量入侵检测系统",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/tasks": {
      "get": {
        "tags": [
          "tasks",
          "tasks"
        ],
        "summary": "获取所有任务状态",
        "description": "获取所有任务状态。\n\n取代旧端点：\n- GET /api/capture/status\n- GET /api/capture/enhanced-status\n- GET /api/detector/status\n- GET /api/attack/status\n- GET /api/auto/status\n- GET /api/dual/stats\n- GET /api/model/train-status",
        "operationId": "list_tasks_api_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/login": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "会话登录",
        "description": "用户名+密码登录，成功后写入会话。",
        "operationId": "login_api_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/change-password": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "修改密码",
        "description": "修改当前用户密码。需提供旧密码验证。",
        "operationId": "change_password_api_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "properties": {
          "old_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Old Password"
          },
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 6,
            "title": "New Password"
          }
        },
        "type": "object",
        "required": [
          "old_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest",
        "description": "修改密码请求。"
      },
      "ChangePasswordResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "title": "ChangePasswordResponse",
        "description": "修改密码响应。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "description": "登录请求。"
      },
      "LoginResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "username": {
            "type": "string",
            "title": "Username",
            "default": ""
          }
        },
        "type": "object",
        "title": "LoginResponse",
        "description": "登录响应。"
      },
...
      "TaskListResponse": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskStatusResponse"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "tasks"
        ],
        "title": "TaskListResponse",
        "description": "任务列表响应。"
      },
      "TaskStatusResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "continuous | timed"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "idle | running | stopping | finished | failed"
          },
          "elapsed": {
            "type": "number",
            "title": "Elapsed",
            "default": 0.0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "default_duration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Duration",
            "description": "限时任务默认时长(秒)"
          },
          "duration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration",
            "description": "限时任务实际启动时长(秒)，前端进度条应用此值"
          }
        },
        "type": "object",
        "required": [
          "name",
          "kind",
          "status"
        ],
        "title": "TaskStatusResponse",
        "description": "单个任务状态响应。\n\nF5 修复：新增 duration 字段，暴露启动时实际使用的窗口时长。\n前端进度条应使用 `duration ?? default_duration` 计算进度。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}